remove ulti class doublon
